Change pwd api

+++ bulkhours/core/install.sh
@@ -1,5 +1,14 @@
 
-cd /content
-rm -rf bulkhours 2> /dev/null && git clone https://github.com/guydegnol/bulkhours.git 2> /dev/null
-echo "git clone https://github.com/guydegnol/bulkhours.git"
-echo "$1" > bulkhours/.safe
+
+if [ -d /content ];
+then
+    rm -rf bulkhours 2> /dev/null && git clone https://github.com/guydegnol/bulkhours.git 2> /dev/null
+    echo "install https://github.com/guydegnol/bulkhours.git package"
+    JSON_FMT='{"pass_code":"%s","package_dir":"%s"}\n'
+    printf "$JSON_FMT" "$1" "/content/bulkhours/" > /content/bulkhours/.safe
+else
+    JSON_FMT='{"pass_code":"%s","package_dir":"%s"}\n'
+    printf "$JSON_FMT" "$1" "/home/guydegnol/projects/bulkhours/" > /home/guydegnol/projects/bulkhours/.safe
+fi
+
+
